Fix strand-specific coverage calculation
`bedtools genomecov` does the wrong thing with `-strand +/-` when
invoked with the `-bg` option: it does not actually respect paired-end
strandedness. The “official” workaround is to filter fragments using
`awk` based on their strand. This commit is instead using a workaround
by computing coverage of R1 only and filtering based on flagged strandedness.